+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [!325][!325] Add `to_uint`, `bits` and `to_int` methods to `int` and `uint`.
 - [!326][!326] Add pattern matching for arrays
+- [!329][!329] Add turbofish (`::<>`) to methods
 
 ### Changed
 
@@ -27,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [!324]: https://gitlab.com/spade-lang/spade/-/merge_requests/324
 [!325]: https://gitlab.com/spade-lang/spade/-/merge_requests/325
 [!326]: https://gitlab.com/spade-lang/spade/-/merge_requests/326
+[!329]: https://gitlab.com/spade-lang/spade/-/merge_requests/329
 
 ## [0.9.0] - 2024-07-04